user can edit and delete only own posts by checking req.userId===Post.creator

+++ controllers/feed.js
@@ -32,6 +32,8 @@ exports.getPosts = (req, res, next) => {
     })
 }
 
+
+
 exports.createPost = (req, res, next) => {
     const errors = validationResult(req);
     if(!errors.isEmpty()){
@@ -131,6 +133,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if(post.creator.toString() !== req.userId){
+                const error = new Error('Not Authorized!')
+                error.statusCode = 403;
+                throw error;
+            }
             if(imageUrl !== post.imageUrl){
                 clearImage(post.imageUrl);
             }
@@ -160,9 +167,21 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if(post.creator.toString() !== req.userId){
+                const error = new Error('Not Authorized!')
+                error.statusCode = 403;
+                throw error;
+            }
             // Check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
         })
         .then(result => {
             res.status(200).json({message: 'Deleted post.'})
@@ -174,6 +193,8 @@ exports.deletePost = (req, res, next) => {
             next(err);
         })
 }
+
+
 
 
 const clearImage = filePath => {
